requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.1.14
gunicorn==20.1.0
Pillow==7.1.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
django-autoslug==1.9.8